Move `Message` implementation for parser `Error`
